requirements.txt
pillow~=8.2.0
cairosvg~=2.5.2
